Update to plc4x/plc4j 0.11 version.

+++ bundles/org.connectorio.addons.binding.canopen/src/main/java/org/connectorio/addons/binding/canopen/internal/config/SocketCANConfiguration.java
@@ -29,7 +29,7 @@ public class SocketCANConfiguration extends CoNodeConfig {
 
   public int retryCount = 3;
 
-  public double throttleReading;
-  public double throttleWriting;
+  public long throttleReading;
+  public long throttleWriting;
 
 }
